ttl: split processor into separate file
Release note: None
